+++ network_as_code/__init__.py
@@ -14,4 +14,4 @@
 
 from .client import NetworkAsCodeClient
 
-__version__ = "2.1.0"
+__version__ = "2.2.0"