+++ network_as_code/api/client.py
@@ -14,11 +14,10 @@
 
 import sys
 from ..api.slice_api import AttachAPI, SliceAPI
-
 from network_as_code.api.qod_api import QodAPI
-
 from .location_api import LocationVerifyAPI, LocationRetrievalAPI
 from .device_status_api import DeviceStatusAPI
+from .congestion_api import CongestionAPI
 
 QOS_BASE_URL_PROD = "https://quality-of-service-on-demand.p-eu.rapidapi.com"
 QOS_RAPID_HOST_PROD = "quality-of-service-on-demand.nokia.rapidapi.com"
@@ -46,6 +45,9 @@ DEVICE_STATUS_BASE_URL_PROD = "https://device-status.p-eu.rapidapi.com"
 DEVICE_STATUS_RAPID_HOST_PROD = "device-status.nokia.rapidapi.com"
 DEVICE_STATUS_BASE_URL_DEV = "https://device-status1.p-eu.rapidapi.com"
     
+CONGESTION_BASE_URL_PROD = "https://congestion-insights.p-eu.rapidapi.com"
+CONGESTION_RAPID_HOST_PROD = "congestion-insights.nokia.rapidapi.com"
+CONGESTION_BASE_URL_DEV = "https://congestion-insights.p-eu.rapidapi.com"
 
 class APIClient:
     """A client for communicating with Network as Code APIs.
@@ -57,16 +59,17 @@ class APIClient:
     """
 
     def __init__(
-        self,
-        token: str,
-        qos_base_url: str = QOS_BASE_URL_PROD,
-        location_verify_base_url: str = LOCATION_VERIFY_BASE_URL_PROD,
-        location_retrieve_base_url: str = LOCATION_RETRIEVE_BASE_URL_PROD,
-        slice_base_url: str = SLICE_BASE_URL_PROD,
-        slice_attach_base_url: str = SLICE_ATTACH_BASE_URL_PROD,
-        device_status_base_url: str = DEVICE_STATUS_BASE_URL_PROD,
-        dev_mode: bool = False,
-        **kwargs,
+            self,
+            token: str,
+            qos_base_url: str = QOS_BASE_URL_PROD,
+            location_verify_base_url: str = LOCATION_VERIFY_BASE_URL_PROD,
+            location_retrieve_base_url: str = LOCATION_RETRIEVE_BASE_URL_PROD,
+            slice_base_url: str = SLICE_BASE_URL_PROD,
+            slice_attach_base_url: str = SLICE_ATTACH_BASE_URL_PROD,
+            device_status_base_url: str = DEVICE_STATUS_BASE_URL_PROD,
+            congestion_base_url: str = CONGESTION_BASE_URL_PROD,
+            dev_mode: bool = False,
+            **kwargs,
     ):
         if dev_mode and qos_base_url == QOS_BASE_URL_PROD:
             qos_base_url = QOS_BASE_URL_DEV
@@ -85,6 +88,9 @@ class APIClient:
 
         if dev_mode and device_status_base_url == DEVICE_STATUS_BASE_URL_PROD:
             device_status_base_url = DEVICE_STATUS_BASE_URL_DEV
+
+        if dev_mode and congestion_base_url == CONGESTION_BASE_URL_PROD:
+            congestion_base_url = CONGESTION_BASE_URL_DEV
 
         self.sessions = QodAPI(
             base_url=qos_base_url,
@@ -147,4 +153,16 @@ class APIClient:
             else slice_attach_base_url.replace("https://", "").replace(
                 "p-eu", "nokia-dev"
             ),
+        )
+
+        self.congestion = CongestionAPI(
+            base_url=congestion_base_url,
+            rapid_key=token,
+            rapid_host=congestion_base_url.replace("https://", "").replace(
+                "p-eu", "nokia"
+            )
+            if not dev_mode
+            else congestion_base_url.replace("https://", "").replace(
+                "p-eu", "nokia-dev"
+            )
         )