+++ network_as_code/api/device_status_api.py
@@ -51,14 +51,14 @@ class DeviceStatusAPI:
         device,
         event_type: str,
         notification_url: str,
-        notification_auth_token: str,
+        notification_auth_token: Optional[str],
         max_number_of_reports: Optional[int] = None,
         subscription_expire_time: Optional[str] = None,
     ):
         assert device.network_access_id != "None"
 
         res = self.client.post(
-            "/event-subscriptions",
+            "/subscriptions",
             json=delete_none(
                 {
                     "subscriptionDetail": {
@@ -74,7 +74,7 @@ class DeviceStatusAPI:
                             else None,
                             "ipv6Address": device.ipv6_address,
                         },
-                        "eventType": event_type,
+                        "type": event_type,
                     },
                     "maxNumberOfReports": max_number_of_reports,
                     "subscriptionExpireTime": subscription_expire_time,
@@ -91,13 +91,38 @@ class DeviceStatusAPI:
         return res.json()
 
     def get_subscription(self, id: str):
-        res = self.client.get(f"/event-subscriptions/{id}")
+        res = self.client.get(f"/subscriptions/{id}")
+
+        error_handler(res)
+
+        return res.json()
+
+    def get_subscriptions(self):
+        res = self.client.get(f"/subscriptions")
 
         error_handler(res)
 
         return res.json()
 
     def delete_subscription(self, id: str):
-        res = self.client.delete(f"/event-subscriptions/{id}")
+        res = self.client.delete(f"/subscriptions/{id}")
 
         error_handler(res)
+
+    def get_connectivity(self, device: dict):
+        res = self.client.post("/connectivity", json={
+            "device": device
+        })
+
+        error_handler(res)
+
+        return res.json()
+
+    def get_roaming(self, device: dict):
+        res = self.client.post("/roaming", json={
+            "device": device
+        })
+
+        error_handler(res)
+
+        return res.json()