+++ network_as_code/namespaces/__init__.py
@@ -17,3 +17,4 @@ from .device import Devices
 from .session import Sessions
 from .slice import Slices
 from .device_status import Connectivity
+from .insights import NetworkInsights