+++ tests/conftest.py
@@ -1,5 +1,6 @@
 
 import os
+import json
 
 import pytest
 
@@ -10,3 +11,6 @@ from network_as_code import NetworkAsCodeClient
 def client() -> NetworkAsCodeClient:
     token = "TEST_TOKEN"
     return NetworkAsCodeClient(token=token)
+
+def to_bytes(json_content: dict) -> bytes:
+    return json.dumps(json_content).encode()