+++ tests/test_device_status_api_mock.py
@@ -1,12 +1,14 @@
 from pytest_httpx import httpx_mock
 import pytest
 import json
+from datetime import datetime
 from httpx import HTTPError
 
 from network_as_code.errors import AuthenticationException, NotFound, ServiceError, APIError
 
 
 from network_as_code.models.device import Device, DeviceIpv4Addr
+from tests.conftest import to_bytes
 
 @pytest.fixture
 def device(client) -> Device:
@@ -23,14 +25,129 @@ def device_with_just_phone_number(client) -> Device:
     device = client.devices.get(phone_number="7777777777")
     return device
 
-def to_bytes(json_content: dict) -> bytes:
-    return json.dumps(json_content).encode()
+def test_updated_device_status_subscription_creation(httpx_mock, client):
+    httpx_mock.add_response(
+        url="https://device-status.p-eu.rapidapi.com/subscriptions",
+        method="POST",
+        json={
+            "subscriptionDetail": {
+                "device": {
+                    "phoneNumber": "123456789",
+                    "networkAccessIdentifier": "123456789@domain.com",
+                    "ipv4Address": {
+                        "publicAddress": "84.125.93.10",
+                        "publicPort": 59765
+                    },
+                    "ipv6Address": "2001:db8:85a3:8d3:1319:8a2e:370:7344"
+                },
+                "type": "org.camaraproject.device-status.v0.roaming-status"
+            },
+            "subscriptionExpireTime": "2023-01-17T13:18:23.682Z",
+            "webhook": {
+                "notificationUrl": "https://application-server.com",
+                "notificationAuthToken": "c8974e592c2fa383d4a3960714"
+            },
+            "subscriptionId": "qs15-h556-rt89-1298",
+            "startsAt": "2024-03-28T12:40:20.398Z",
+            "expiresAt": "2024-03-28T12:40:20.398Z"
+        },
+        match_content=to_bytes(
+            {
+                "subscriptionDetail": {
+                    "device": {
+                        "phoneNumber": "123456789",
+                        "networkAccessIdentifier": "123456789@domain.com",
+                        "ipv4Address": {
+                            "publicAddress": "84.125.93.10",
+                            "publicPort": 59765
+                        },
+                        "ipv6Address": "2001:db8:85a3:8d3:1319:8a2e:370:7344"
+                    },
+                    "type": "org.camaraproject.device-status.v0.roaming-status"
+                },
+                "subscriptionExpireTime": "2023-01-17T13:18:23.682Z",
+                "webhook": {
+                    "notificationUrl": "https://application-server.com",
+                    "notificationAuthToken": "c8974e592c2fa383d4a3960714"
+                }
+            }
+        )
+    )
+
+    device = client.devices.get("123456789@domain.com", phone_number="123456789", ipv4_address=DeviceIpv4Addr(public_address="84.125.93.10", public_port=59765), ipv6_address="2001:db8:85a3:8d3:1319:8a2e:370:7344")
+
+    subscription = client.connectivity.subscribe(
+        device=device,
+        event_type="org.camaraproject.device-status.v0.roaming-status",
+        subscription_expire_time="2023-01-17T13:18:23.682Z",
+        notification_url="https://application-server.com",
+        notification_auth_token="c8974e592c2fa383d4a3960714"
+    )
+
+def test_subscribing_using_datetime(httpx_mock, client):
+    httpx_mock.add_response(
+        url="https://device-status.p-eu.rapidapi.com/subscriptions",
+        method="POST",
+        json={
+            "subscriptionDetail": {
+                "device": {
+                    "phoneNumber": "123456789",
+                    "networkAccessIdentifier": "123456789@domain.com",
+                    "ipv4Address": {
+                        "publicAddress": "84.125.93.10",
+                        "publicPort": 59765
+                    },
+                    "ipv6Address": "2001:db8:85a3:8d3:1319:8a2e:370:7344"
+                },
+                "type": "org.camaraproject.device-status.v0.roaming-status"
+            },
+            "subscriptionExpireTime": "2023-01-17T13:18:23.682000+00:00",
+            "webhook": {
+                "notificationUrl": "https://application-server.com",
+                "notificationAuthToken": "c8974e592c2fa383d4a3960714"
+            },
+            "subscriptionId": "qs15-h556-rt89-1298",
+            "startsAt": "2024-03-28T12:40:20.398Z",
+            "expiresAt": "2024-03-28T12:40:20.398Z"
+        },
+        match_content=to_bytes(
+            {
+                "subscriptionDetail": {
+                    "device": {
+                        "phoneNumber": "123456789",
+                        "networkAccessIdentifier": "123456789@domain.com",
+                        "ipv4Address": {
+                            "publicAddress": "84.125.93.10",
+                            "publicPort": 59765
+                        },
+                        "ipv6Address": "2001:db8:85a3:8d3:1319:8a2e:370:7344"
+                    },
+                    "type": "org.camaraproject.device-status.v0.roaming-status"
+                },
+                "subscriptionExpireTime": "2023-01-17T13:18:23.682000+00:00",
+                "webhook": {
+                    "notificationUrl": "https://application-server.com",
+                    "notificationAuthToken": "c8974e592c2fa383d4a3960714"
+                }
+            }
+        )
+    )
+
+    device = client.devices.get("123456789@domain.com", phone_number="123456789", ipv4_address=DeviceIpv4Addr(public_address="84.125.93.10", public_port=59765), ipv6_address="2001:db8:85a3:8d3:1319:8a2e:370:7344")
+
+    subscription = client.connectivity.subscribe(
+        device=device,
+        event_type="org.camaraproject.device-status.v0.roaming-status",
+        subscription_expire_time=datetime.fromisoformat("2023-01-17T13:18:23.682+00:00"),
+        notification_url="https://application-server.com",
+        notification_auth_token="c8974e592c2fa383d4a3960714"
+    )
 
 def test_device_status_creation_minimal_parameters(httpx_mock, device, client):
     httpx_mock.add_response(
         method="POST",
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
         },
         match_content=to_bytes({
             "subscriptionDetail": {
@@ -42,9 +159,8 @@ def test_device_status_creation_minimal_parameters(httpx_mock, device, client):
                         "publicPort": 80
                     },
                 },
-                "eventType": "CONNECTIVITY"
+                "type": "CONNECTIVITY"
             },
-            "maxNumberOfReports": 1,
             "webhook": {
                 "notificationUrl": "https://localhost:9090/notify",
                 "notificationAuthToken": "my_auth_token"
@@ -52,13 +168,13 @@ def test_device_status_creation_minimal_parameters(httpx_mock, device, client):
         })
     )
 
-    subscription = client.connectivity.subscribe("CONNECTIVITY", 1, "https://localhost:9090/notify", device, "my_auth_token")
+    subscription = client.connectivity.subscribe(event_type="CONNECTIVITY", notification_url="https://localhost:9090/notify", device=device, notification_auth_token="my_auth_token")
 
 def test_device_status_creation_minimal_parameters_minimal_ipv4(httpx_mock, device_with_just_public_ipv4, client):
     httpx_mock.add_response(
         method="POST",
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
         },
         match_content=to_bytes({
             "subscriptionDetail": {
@@ -68,7 +184,7 @@ def test_device_status_creation_minimal_parameters_minimal_ipv4(httpx_mock, devi
                         "publicAddress": "1.1.1.2"
                     }
                 },
-                "eventType": "CONNECTIVITY"
+                "type": "CONNECTIVITY"
             },
             "maxNumberOfReports": 1,
             "webhook": {
@@ -78,20 +194,20 @@ def test_device_status_creation_minimal_parameters_minimal_ipv4(httpx_mock, devi
         })
     )
 
-    subscription = client.connectivity.subscribe("CONNECTIVITY", 1, "https://localhost:9090/notify", device_with_just_public_ipv4, "my_auth_token")
+    subscription = client.connectivity.subscribe(event_type="CONNECTIVITY", notification_url="https://localhost:9090/notify", device=device_with_just_public_ipv4, notification_auth_token="my_auth_token", max_num_of_reports=1)
 
 def test_device_status_creation_minimal_parameters_only_phone_number(httpx_mock, device_with_just_phone_number, client):
     httpx_mock.add_response(
         method="POST",
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
         },
         match_content=to_bytes({
             "subscriptionDetail": {
                 "device": {
                     "phoneNumber": "7777777777"
                 },
-                "eventType": "CONNECTIVITY"
+                "type": "CONNECTIVITY"
             },
             "maxNumberOfReports": 1,
             "webhook": {
@@ -101,12 +217,12 @@ def test_device_status_creation_minimal_parameters_only_phone_number(httpx_mock,
         })
     )
 
-    subscription = client.connectivity.subscribe("CONNECTIVITY", 1, "https://localhost:9090/notify", device_with_just_phone_number, "my_auth_token")
+    subscription = client.connectivity.subscribe(event_type="CONNECTIVITY", notification_url="https://localhost:9090/notify", device=device_with_just_phone_number, notification_auth_token="my_auth_token", max_num_of_reports=1)
 
 def test_device_status_creation_with_optional_parameters(httpx_mock, device, client):
     httpx_mock.add_response(
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
         },
         match_content=to_bytes({
             "subscriptionDetail": {
@@ -118,7 +234,7 @@ def test_device_status_creation_with_optional_parameters(httpx_mock, device, cli
                         "publicPort": 80
                     },
                 },
-                "eventType": "CONNECTIVITY"
+                "type": "CONNECTIVITY"
             },
             "maxNumberOfReports": 1,
             "subscriptionExpireTime": "2023-08-31",
@@ -129,13 +245,13 @@ def test_device_status_creation_with_optional_parameters(httpx_mock, device, cli
         })
     )
     
-    subscription = client.connectivity.subscribe("CONNECTIVITY", 1, "https://localhost:9090/notify", device, "my_auth_token", subscription_expire_time="2023-08-31")
+    subscription = client.connectivity.subscribe(event_type="CONNECTIVITY", notification_url="https://localhost:9090/notify", device=device, notification_auth_token="my_auth_token", subscription_expire_time="2023-08-31", max_num_of_reports=1)
 
 def test_device_status_creation_with_roaming_status(httpx_mock, device, client):
     httpx_mock.add_response(
         method="POST",
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
         },
         match_content=to_bytes({
             "subscriptionDetail": {
@@ -147,7 +263,7 @@ def test_device_status_creation_with_roaming_status(httpx_mock, device, client):
                         "publicPort": 80
                     },
                 },
-                "eventType": "ROAMING_STATUS"
+                "type": "ROAMING_STATUS"
             },
             "maxNumberOfReports": 1,
             "webhook": {
@@ -157,13 +273,14 @@ def test_device_status_creation_with_roaming_status(httpx_mock, device, client):
         })
     )
 
-    subscription = client.connectivity.subscribe("ROAMING_STATUS", 1, "https://localhost:9090/notify", device, "my_auth_token")
+    subscription = client.connectivity.subscribe(event_type="ROAMING_STATUS", notification_url="https://localhost:9090/notify", device=device, notification_auth_token="my_auth_token", max_num_of_reports=1)
 
 def test_getting_device_status_subscription(httpx_mock, device, client):
     httpx_mock.add_response(
+        url="https://device-status.p-eu.rapidapi.com/subscriptions/test-subscription",
         method="GET",
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
             "subscriptionDetail": {
                 "device": {
                     "networkAccessIdentifier": "testuser@open5glab.net",
@@ -171,7 +288,7 @@ def test_getting_device_status_subscription(httpx_mock, device, client):
                         "publicAddress": "1.1.1.2",
                     },
                 },
-                "eventType": "CONNECTIVITY"
+                "type": "CONNECTIVITY"
             },
             "maxNumberOfReports": 1,
             "webhook": {
@@ -189,7 +306,7 @@ def test_deleting_device_status_subscription(httpx_mock, device, client):
     httpx_mock.add_response(
         method="GET",
         json={
-            "eventSubscriptionId": "test-subscription",
+            "subscriptionId": "test-subscription",
             "subscriptionDetail": {
                 "device": {
                     "networkAccessIdentifier": "testuser@open5glab.net",
@@ -197,7 +314,7 @@ def test_deleting_device_status_subscription(httpx_mock, device, client):
                         "publicAddress": "1.1.1.2"
                     },
                 },
-                "eventType": "CONNECTIVITY"
+                "type": "CONNECTIVITY"
             },
             "maxNumberOfReports": 1,
             "webhook": {
@@ -211,11 +328,153 @@ def test_deleting_device_status_subscription(httpx_mock, device, client):
     subscription = client.connectivity.get_subscription("test-subscription")
 
     httpx_mock.add_response(
+        url="https://device-status.p-eu.rapidapi.com/subscriptions/test-subscription",
         method="DELETE",
     )
 
     subscription.delete()
 
+def test_get_subscriptions(httpx_mock, device, client):
+    httpx_mock.add_response(
+        method="GET",
+        url="https://device-status.p-eu.rapidapi.com/subscriptions",
+        json=[
+            {
+                "subscriptionDetail": {
+                    "device": {
+                        "networkAccessIdentifier": "testuser@testcsp.net"
+                    },
+                    "type": "org.camaraproject.device-status.v0.connectivity-data"
+                },
+                "maxNumberOfReports": 1,
+                "webhook": {
+                    "notificationUrl": "https://example.com"
+                },
+                "subscriptionId": "34e9e3ee-e281-4f47-bbc2-2431e6abbef0",
+                "eventSubscriptionId": "34e9e3ee-e281-4f47-bbc2-2431e6abbef0",
+                "startsAt": "2024-04-09T11:14:50.254312Z",
+                "expiresAt": "2024-04-10T14:13:29.766268"
+            },
+            {
+                "subscriptionDetail": {
+                    "device": {
+                        "networkAccessIdentifier": "testuser@testcsp.net"
+                    },
+                    "type": "org.camaraproject.device-status.v0.connectivity-data"
+                },
+                "maxNumberOfReports": 1,
+                "webhook": {
+                    "notificationUrl": "https://example.com"
+                },
+                "subscriptionId": "51b24d1a-26ae-4c9d-b114-2086da958c50",
+                "eventSubscriptionId": "51b24d1a-26ae-4c9d-b114-2086da958c50",
+                "startsAt": "2024-04-09T11:21:22.871187Z",
+                "expiresAt": "2024-04-10T14:13:29Z"
+            },
+            {
+                "subscriptionDetail": {
+                    "device": {
+                        "networkAccessIdentifier": "sdk-integration@testcsp.net",
+                        "ipv4Address": {
+                            "publicAddress": "1.1.1.2",
+                            "privateAddress": "1.1.1.2",
+                            "publicPort": 80
+                        }
+                    },
+                    "type": "org.camaraproject.device-status.v0.roaming-status",
+                    "eventType": "ROAMING_STATUS"
+                },
+                "maxNumberOfReports": 1,
+                "webhook": {
+                    "notificationUrl": "http://192.0.2.0:8080/",
+                    "notificationAuthToken": "c8974e592c2fa383d4a3960714"
+                },
+                "subscriptionId": "815e6da4-813d-4111-987d-5e6036aaa410",
+                "eventSubscriptionId": "815e6da4-813d-4111-987d-5e6036aaa410",
+                "startsAt": "2024-04-05T14:29:56.792078Z"
+            },
+            {
+                "subscriptionDetail": {
+                    "device": {
+                        "networkAccessIdentifier": "sdk-integration@testcsp.net",
+                        "ipv4Address": {
+                            "publicAddress": "1.1.1.2",
+                            "privateAddress": "1.1.1.2",
+                            "publicPort": 80
+                        }
+                    },
+                    "type": "org.camaraproject.device-status.v0.connectivity-data"
+                },
+                "maxNumberOfReports": 1,
+                "webhook": {
+                    "notificationUrl": "http://192.0.2.0:8080/",
+                    "notificationAuthToken": "c8974e592c2fa383d4a3960714"
+                },
+                "subscriptionId": "f6b03776-5e0f-4dbc-abce-30a916f94ad0",
+                "eventSubscriptionId": "f6b03776-5e0f-4dbc-abce-30a916f94ad0",
+                "startsAt": "2024-04-09T11:11:27.052869Z",
+                "expiresAt": "2025-04-08T14:13:29.766268"
+            }
+        ]
+    )
+
+    subscriptions = client.connectivity.get_subscriptions()
+
+    assert len(subscriptions) > 0
+
+    for subscription in subscriptions:
+        assert subscription.id
+        assert subscription.device
+
+def test_poll_connectivity(httpx_mock, device, client):
+    httpx_mock.add_response(
+        method="POST",
+        url="https://device-status.p-eu.rapidapi.com/connectivity",
+        json={
+            "connectivityStatus": "CONNECTED_DATA"
+        },
+        match_content=to_bytes({
+            "device": {
+                "networkAccessIdentifier": "testuser@open5glab.net",
+                "ipv4Address": {
+                    "publicAddress": "1.1.1.2",
+                    "privateAddress": "1.1.1.2",
+                    "publicPort": 80
+                }
+            }
+        })
+    )
+
+    status = device.get_connectivity()
+
+    assert status == "CONNECTED_DATA"
+
+def test_poll_roaming(httpx_mock, device, client):
+    httpx_mock.add_response(
+        method="POST",
+        url="https://device-status.p-eu.rapidapi.com/roaming",
+        json={
+            "roaming": True,
+            "countryCode": 358,
+            "countryName": ["Finland"]
+        },
+        match_content=to_bytes({
+            "device": {
+                "networkAccessIdentifier": "testuser@open5glab.net",
+                "ipv4Address": {
+                    "publicAddress": "1.1.1.2",
+                    "privateAddress": "1.1.1.2",
+                    "publicPort": 80
+                }
+            }
+        })
+    )
+
+    status = device.get_roaming()
+
+    assert status.roaming
+    assert status.country_code == 358
+    assert status.country_name == ["Finland"]
 
 def test_subscribe_authentication_exception(httpx_mock, device, client):
     httpx_mock.add_response(