+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 # Changelog
 
+## Version 2.2
+
+Breaking changes:
+- None!
+
+Changes:
+- Updated API bindings for Device Status and integrated new connectivity and roaming status polling
+- Integrated new Congestion Insights API for querying and predicting network congestion
+
 ## Version 2.1
 
 Breaking changes: